% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_stats.R
\name{order_rchisq}
\alias{order_rchisq}
\title{Gets random order statistics from a chi-square distribution}
\usage{
order_rchisq(draw_size, df, k, n)
}
\arguments{
\item{draw_size}{- The size of the output sample}

\item{df}{- The degrees of the chi-square distribution}

\item{k}{- The Kth smallest value from a sample}

\item{n}{- The size of the sample to compute the order statistic from}
}
\value{
A vector of random order statistic variables from a chi-square distribution
}
\description{
Gets random order statistics from a chi-square distribution
}
\examples{
order_rchisq(10, 1, 100, 10000)
}

