% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_stats.R
\name{order_rlogis}
\alias{order_rlogis}
\title{Gets random order statistics from a logistic distribution}
\usage{
order_rlogis(draw_size, location, scale, k, n)
}
\arguments{
\item{draw_size}{- The size of the output sample}

\item{location}{- The location parameter in the logistic distribution}

\item{scale}{- The scale parameter in the logistic distribution}

\item{k}{- The Kth smallest value from a sample}

\item{n}{- The size of the sample to compute the order statistic from}
}
\value{
A vector of random order statistic variables from a logistic distribution
}
\description{
Gets random order statistics from a logistic distribution
}
\examples{
order_rlogis(10, 0, 1, 100, 10000)
}

