% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_stats.R
\name{order_rnorm}
\alias{order_rnorm}
\title{Gets random order statistics from a normal distribution}
\usage{
order_rnorm(draw_size = 1, mean = 0, sd = 1, k = 1, n = 1)
}
\arguments{
\item{draw_size}{- The size of the output sample}

\item{mean}{- The mean of the normal distribution}

\item{sd}{- The standard deviation of the normal distribution}

\item{k}{- The Kth smallest value from a sample}

\item{n}{- The size of the sample to compute the order statistic from}
}
\value{
A vector of random order statistic variables from a normal distribution
}
\description{
Gets random order statistics from a normal distribution
}
\examples{
order_rnorm(10, 0, 1, 100, 10000)
}

